split out sampling compressor
